Moved config mock into testhelpers area.
